a few more minor cleanups and reordering
